feature: support in-memory database
In some cases, static user data is sufficient and changes to the users
do not need to be persisted. This change introduces a special file path
`:memory:` (similar to sqlite3) that allows the user to specify that the
database is in-memory only. Changes will not be persisted on the file
system.