fix: correct length check of profile api key
API keys are created as a random string in the range of 64 to 72
characters. This range is enforced in the key suggested by the
`/profile/api-keys/new` endpoint (see javascript function [`pD()`][js]),
as well as in the database function [`AddAPIKey()`][database], which
generates and validates a key if none is passed. However, the HTTP
handler that checks the key passed from the UI uses and incorrect range,
generating the error reported in
https://github.com/greenpau/caddy-security/issues/395. Correct the
length range.

NOTE: This issue probably went unnoticed in
https://github.com/greenpau/go-authcrunch/commit/c5d0307a6cc2d61cfb372dce64bf477943189f8d,
when the length range was changed.

[js]: https://github.com/greenpau/go-authcrunch/blob/main/pkg/authn/ui/profile/assets/index-DOtSzC14.js?h=1#L1468
[database]: https://github.com/greenpau/go-authcrunch/blob/main/pkg/identity/database.go#L601